refactor(api): move response records into the endpoint files
Deletes the Contracts/ folder in both API services; each response record
now lives in the endpoint file that serves it - PatientEndpoints.cs holds
PatientResponse, SampleResponse and SpecimenResponse, HealthEndpoints.cs
holds HealthResponse.

A route and its wire shape are one thing and change together, so reading
the route shows the contract. This is the one place the one-type-per-file
convention is deliberately relaxed.

No behaviour change: the records, their fields and the JSON are
identical. Only the namespace moves, from Web.Contracts to Web.Endpoints.
