--- src/BiobankApi/BiobankApi.Web/Contracts/HealthResponse.cs
@@ -1,4 +1,0 @@
-namespace BiobankApi.Web.Contracts;
-
-/// <summary>Response shape for <c>GET /health</c>.</summary>
-public sealed record HealthResponse(string Status);
