--- src/SequencingApi/SequencingApi.Web/Contracts/HealthResponse.cs
@@ -1,4 +1,0 @@
-namespace SequencingApi.Web.Contracts;
-
-/// <summary>Response shape for <c>GET /health</c>.</summary>
-public sealed record HealthResponse(string Status);
